Working

+++ app/Http/Controllers/JobController.php
@@ -5,23 +5,47 @@ namespace App\Http\Controllers;
 use App\Http\Requests\CreateJobRequest;
 use App\Models\Job;
 use Illuminate\Http\Request;
+use Illuminate\Http\Exceptions\HttpResponseException;
 
 class JobController extends Controller
 {
 
-    public function index()
+    public function all_jobs(){
+
+        return response()->json([
+            'success'   => true,
+            'message'   => 'All public jobs',
+            'data'      => Job::all(),
+        ]);
+        
+    }
+
+    public function index(Request $request)
     {
+
+        $user = $request->user();
+        $jobs = Job::where('user_id', $user->id)->get();
+
         return response()->json([
             'success'   => true,
             'message'   => 'All jobs',
-            'data'      => Job::all(),
+            'data'      => $jobs,
         ]);
     }
 
     public function store(CreateJobRequest $request)
     {
+
+        $user = $request->user();
+
+        
+        $job_data = $request->all();
+        $job_data['user_id'] = $user->id;
+
+        // dd($job_data);
+
         // Create the job
-        $job = Job::create($request->all());
+        $job = Job::create($job_data);
 
         return response()->json([
             'success'   => true,
@@ -32,17 +56,44 @@ class JobController extends Controller
 
     public function get_job_by_id($id)
     {
+
+        $job = Job::find($id);
+        if(!$job) {
+            throw new HttpResponseException(response()->json([
+                'success'   => false,
+                'message'   => 'Job not found',
+            ], 400));
+        }
+
         return response()->json([
             'success'   => true,
             'message'   => 'Successfully retrieved job',
-            'data'      => Job::find($id),
+            'data'      => $job,
         ]);
     }
 
     // update job
     public function update_job_by_id($id, CreateJobRequest $request)
     {
+
+        $user = $request->user();
+
         $job = Job::find($id);
+
+        if(!$job) {
+            throw new HttpResponseException(response()->json([
+                'success'   => false,
+                'message'   => 'Job not found',
+            ], 400));
+        }
+
+        if($job->user_id != $user->id) {
+            throw new HttpResponseException(response()->json([
+                'success'   => false,
+                'message'   => 'Access denied',
+            ], 400));
+        }
+
         $job->update($request->all());
         return response()->json([
             'success'   => true,
@@ -56,9 +107,23 @@ class JobController extends Controller
     {
 
         $user = $request->user();
+        $job = Job::find($id);
+
+        if(!$job) {
+            throw new HttpResponseException(response()->json([
+                'success'   => false,
+                'message'   => 'Job not found',
+            ], 400));
+        }
+
+        if($job->user_id != $user->id) {
+            throw new HttpResponseException(response()->json([
+                'success'   => false,
+                'message'   => 'Access denied',
+            ], 400));
+        }
 
         // Delete the job
-        $job = Job::find($id);
         $job->delete();
         return response()->json([
             'success'   => true,
